==> BOOK/01-construct_new_data.ipynb <==
import numpy as np
import pandas as pd

user_book = pd.read_csv("./Book_data/old_data/Ratings.csv")   # http://www2.informatik.uni-freiburg.de/~cziegler/BX/

user_book = user_book[(user_book["Book-Rating"] > 5) ]
user_book

user_ID = set()
book_ID = set()
for i in  user_book['User-ID']:
    user_ID.add(i)
for i in  user_book['ISBN']:
    book_ID.add(i)
print(len(user_ID), len(book_ID))
print("平均每个人交互的次数:",len(user_book )/len(user_ID))

books = pd.read_csv("./Book_data/old_data/Books.csv")
users = pd.read_csv("./Book_data/old_data/Users.csv")

#  找出与用户交互过的书
books = books[books["ISBN"].isin(book_ID)]
users = users[users["User-ID"].isin(user_ID)]
print(len(users),len(books))
# len(books)  <  用户看过的   一部分用户交互过的书在books里没有，进一步处理

# 用户交互过的书
user_book = user_book[user_book["ISBN"].isin(books["ISBN"])]

user_ID = set()
book_ID = set()
for i in  user_book['User-ID']:
    user_ID.add(i)
for i in  user_book['ISBN']:
    book_ID.add(i)
print("用户数：%d, 书集数：%d, 交互数：%d" %(len(user_ID), len(book_ID), len(user_book)))
print("平均每个人交互的次数:",len(user_book ) /len(user_ID))
# 用户数：61288, 书集数：129014, 交互数：321896
# 平均每个人交互的次数: 5.252186398642475

users = users[users["User-ID"].isin(user_ID)]
len(users)

Ratings = "./Book_data/new_data/Ratings.csv"
Books = "./Book_data/new_data/Books.csv"
Users = "./Book_data/new_data/Users.csv"
user_book.to_csv(Ratings,sep=',', columns=None, header=True, index=False, mode='w', encoding=None)
books.to_csv(Books,sep=',', columns=None, header=True, index=False, mode='w', encoding=None)
users.to_csv(Users,sep=',', columns=None, header=True, index=False, mode='w', encoding=None)

pd.read_csv("./Book_data/new_data/Books.csv")
pd.read_csv("./Book_data/new_data/Users.csv")
