==> BOOK/02-user_book_ID.ipynb <==
import numpy as np
import pandas as pd

Books = pd.read_csv("./Book_data/new_data/Books.csv")
Users = pd.read_csv("./Book_data/new_data/Users.csv")
Ratings = pd.read_csv("./Book_data/new_data/Ratings.csv")

#  将文件的ID以字典的形式换成数字
def replaceID_Num(rel_data, colum):
    id_n = {}
    count = 0
    for id in rel_data[colum]:
        if id in id_n:
            continue
        else:
            id_n[id] =count
            count += 1
    return id_n

# 找出某一列类型里元素重复出现的次数
def ID_Num(rel_data, u_data, colum):
    id_num= {}   # colum中元素重复出现的次数
    id_n = {}    # 为colum中元素重新定义下标，即为字典的值
    count = 0
    for id in u_data[colum]:
        id_num[id]=1
    for id in rel_data[colum]:
        if id in id_n:
            id_num[id] += 1
        else:
            id_n[id] =count
            count += 1
    return id_n,id_num

_, user_num = ID_Num(Ratings, Users, "User-ID")

rel = 0
Users_ID = set()
for i in user_num:
    if user_num[i] >= 5:
        rel += user_num[i]
        Users_ID.add(i)
print("用户数%d， 总交互数：%d" %(len(Users_ID), rel))

Ratings

Ratings_5 = Ratings[Ratings["User-ID"].isin(Users_ID)]
Ratings_5

Books_ID = set()
for i in  Ratings_5['ISBN']:
    Books_ID.add(i)
print(len(Users_ID), len(Books_ID))
print("平均每个人交互的次数:",len(Ratings_5 )/len(Users_ID))

books = pd.read_csv("./Book_data/new_data/Books.csv")
users = pd.read_csv("./Book_data/new_data/Users.csv")
print(len(users),len(books))

users = users[users["User-ID"].isin(Users_ID)]
books = books[books["ISBN"].isin(Books_ID)]
print(len(users),len(books))

Users_n,Books_n,Locations_n=  replaceID_Num(users, "User-ID"),replaceID_Num(books,"ISBN"),replaceID_Num(users,"Location")
Book_Author_n,Year_Publication_n,Publisher_n=  replaceID_Num(books, "Book-Author"),replaceID_Num(books,"Year-Of-Publication"),replaceID_Num(books,"Publisher")

print("length_all:users:%d, books:%d, locations:%d "  %(len(Users_n), len(Books_n), len(Locations_n)))
print("length_all:Book_Author:%d, Year_Publication:%d, Publisher:%d "  %(len(Book_Author_n), len(Year_Publication_n), len(Publisher_n)))

# 将目标文件的数据替换为对应的数字并写入文件
Users_filename = "./Book_data/new_data/Users_5.csv"
Ratings_filename = "./Book_data/new_data/Ratings_5.csv"
Books_filename = "./Book_data/new_data/Books_5.csv"

# Users
columns = []
f = open(Users_filename, "w")
for each in users.columns:
    columns.append(each)
f.write("User-ID" + ' ' + "Location" + ' ' + "Age" + '\n')  
for ID1,ID2,ID3 in zip(users[columns[0]],users[columns[1]],users[columns[2]]):
    f.write(str(Users_n[ID1]) + ' ' + str(Locations_n[ID2]) + ' ' + str(ID3) + '\n')   
f.close()

books.head(1)

# Books
columns = []
f = open(Books_filename, "w")
for each in books.columns:
    columns.append(each)
f.write("ISBN" + ' ' + "Book-Author" + ' ' + "Year-Of-Publication" +  ' ' +  "Publisher" + '\n')  
for ID1,ID2,ID3,ID4 in zip(books[columns[0]],books[columns[2]],books[columns[3]], books[columns[4]]):
    f.write(str(Books_n[ID1]) + ' ' + str(Book_Author_n[ID2]) + ' ' + str(int(ID3)) + ' ' + str(Publisher_n[ID4]) +'\n')   
f.close()

# Books_n , Book_Author_n,Year_Publication_n,Publisher_n

Ratings_5.head(10)

# Rating
columns = []
f = open(Ratings_filename, "w")
for each in Ratings_5.columns:
    columns.append(each)
f.write("User-ID" + ' ' + "ISBN" + ' ' + "Book-Rating" + '\n')  
for ID1,ID2,ID3 in zip(Ratings_5[columns[0]],Ratings_5[columns[1]],Ratings_5[columns[2]]):
    f.write(str(Users_n[ID1]) + ' ' + str(Books_n[ID2])  + ' '+ str(ID3) +'\n')   
f.close()


# 将目标文件的数据替换为对应的数字并写入文件
def replaceData_Num(data, new_filename, new_ID1,new_ID2):   
    columns = []
    f  = open(new_filename, "w")
    for each in data.columns:
        columns.append(each)
    if len(columns) == 2:
        for ID1,ID2 in zip(data[columns[0]],data[columns[1]]):
    #     print(user_n[ID1],video_n[ID2])
            f.write(str(new_ID1[ID1]) + ' ' + str(new_ID2[ID2]) + '\n')
    else:
         for ID1 in data[columns[0]]:
            f.write(str(new_ID1[ID1]) + '\n')         
    f.close()

Books = pd.read_csv("./Book_data/new_data/Books_5.csv", sep = ' ')
Books
