==> BOOK/04-groupMember.ipynb <==
import numpy as np
import pandas as pd

Ratings = pd.read_csv("./Book_data/new_data/Ratings_5.csv", sep = ' ')

Ratings = Ratings.sort_values(by="User-ID")
Ratings

import random
user = set()
for i in Ratings['User-ID']:
    user.add(i)
random.shuffle(list(user))
user = list(user)
len(user)

#  每个组3-5人
groupID = []
groupMemberOne = []
groupMember = []
j = np.random.randint(3,6)
C_groupID = 0
C_len =0
for i in range(len(user)):
    C_len += j
    if j > 1:  
        j -= 1
        groupMemberOne.append(user[i] )
    elif j == 1:
        groupMemberOne.append(user[i] )
        groupMember.append(groupMemberOne)
        groupMemberOne =[]
        j = np.random.randint(3,6)
#         print(j)
        groupID.append(C_groupID)
        C_groupID += 1
    
if C_len > len(user) and groupMemberOne is not None:
    groupMember.append(groupMemberOne)
    C_groupID += 1
    groupID.append(C_groupID)

print("组的个数：",len(groupMember))
groupMember[-20:]

filename = './Book_data/KA_Book/moocgroupMember_u_v.csv'
GroupID = 0
count = 0
with open(filename,'w') as object:
    for i in groupMember:
        object.write(str(GroupID)+" ")
        count=0
        for j in i:
            c = len(i)
            count+=1
            if count < c:
                object.write(str(j))
                object.write(',')
            else:
                object.write(str(j))
        GroupID+=1
        object.write('\n')

groupMembe = pd.read_csv("./Book_data/KA_Book/moocgroupMember_u_v.csv",header=None,sep =' ')
groupMembe
