==> BOOK/06-KGCN_kg_data.ipynb <==
import numpy as np
import pandas as pd

book  = pd.read_csv("./Book_data/new_data/Books_5.csv", sep = ' ')
book

book.sort_values(by = "Publisher")

year = set()
for i in book['Year-Of-Publication']:
    year.add(i)
year = sorted(year)
print((year))

year_new = {}
count = 0
for i in year:
    year_new[i] = count
    count += 1
year_new

#  ISBN:111336    Book-Author:47288  year:95 Publisher:9053
filename = './Book_data/KA_Book/mooc_kg_video.csv'
with open(filename,'w') as object:
    for i,j in zip(book["ISBN"],book['Book-Author']):
        object.write(str(i) + "\t0\t" + str(j+111336))        # book - 0 - Author
        object.write('\n')
    for i,j in zip(book["ISBN"],book['Year-Of-Publication']):
        if year_new[j] <= 21:
            object.write(str(i) + "\t1\t" + str(year_new[j] + 111336 + 47288))        # book - 1 - year
            object.write('\n')
        elif year_new[j] > 21 and year_new[j] <= 41:
            object.write(str(i) + "\t2\t" + str(year_new[j] + 111336 + 47288))        # book - 2 - year
            object.write('\n')
        elif year_new[j] > 41 and year_new[j] <= 61:
            object.write(str(i) + "\t3\t" + str(year_new[j] + 111336 + 47288))        # book - 3 - year
            object.write('\n')
        elif year_new[j] > 61 and year_new[j] <= 81:
            object.write(str(i) + "\t4\t" + str(year_new[j] + 111336 + 47288))        # book - 4 - year
            object.write('\n')
        else:
            object.write(str(i) + "\t5\t" + str(year_new[j] + 111336 + 47288))        # book - 5 - year
            object.write('\n')
    for i,j in zip(book["ISBN"],book['Publisher']):
        object.write(str(i) + "\t6\t" + str(j + 111336 + 47288 + 95))        # book - 6 - Publisher
        object.write('\n')
